Refactor retrive type
